refactor: extract challenge logic
